*07* Add Bonus Time
*Enemy and EnemyShootBody(implements from HasBullet) are generate from
the same place

** set healthy = true when = flase -> Bonus Time (gen 2 more enemy but
when recieve = 5 ->healthy = true again)
** showDamage works now!! but clearRect is not good :(
**Try to short code

+++ f2/spw/GameReporter.java
@@ -9,6 +9,6 @@ public interface GameReporter {
 	// get Current X,Y axis of spaceship
 	int getCurrentXOfSS();
 	int getCurrentYOfSS();
-	int getDamage();
+	long getDamage();
 
 }
